added kwargs to Network init for user-defined Network members. renamed router => routers to reflect naming convention.
